+++ Changelog.md
@@ -1,2 +1,8 @@
+## Version 1.1.0
+
+- Added support for custom device and custom firmware #23
+- Added support for physical serialport on Windows #26
+
 ## Version 1.0.0
+
 - Initial public release